Code clean up. Improved recovery from errors and disconnections.
