+++ src/constants.ts
@@ -1,0 +1,1 @@
+export const FEE_PER_KB = 2000000;
