+++ src/pages/TokenList.tsx
@@ -1,5 +1,5 @@
 import { useEffect, useState } from "react";
-import { blockchain, fetchDeployments } from "../blockchain";
+import { changeToken } from "../blockchain";
 import {
   Box,
   Button,
@@ -27,10 +27,11 @@ import { LuRefreshCw } from "react-icons/lu";
 import { Link, useNavigate, useParams } from "react-router-dom";
 import Pagination from "../Pagination";
 import { reverseRef } from "../utils";
-import { miningStatus, selectedContract } from "../signals";
+import { miningEnabled, miningStatus, selectedContract } from "../signals";
 import ShortRef from "../ShortRef";
 import miner from "../miner";
 import { addMessage } from "../message";
+import { fetchDeployments } from "../deployments";
 
 function TokenRow({ token }: { token: ContractGroup }) {
   const { target, reward, contractRef } = token.contracts[0];
@@ -51,11 +52,12 @@ function TokenRow({ token }: { token: ContractGroup }) {
     );
 
     selectedContract.value = ref;
-    blockchain.changeToken(ref);
+    changeToken(ref);
     if (miningStatus.value !== "ready") {
       addMessage({ type: "stop" });
     }
     miner.stop();
+    miningEnabled.value = false;
     navigate("/");
   };
